Crypt::Passwd::XS 0.506
Case 47687: Fixed multiple segfaults in the XS hooks to the underlying crypt
implementations when undef was passed in as either plaintext or salt. The test
suite was updated to check for these errors and to improve the output when
errors are encountered. DES crypt C implementation was also fixed to produce
rehashable results when single character salts are used.

Tidy lib/Crypt/Passwd/XS.pm